hrm_py/hrm_py/requirements.txt
Django>=4.2
djangorestframework>=3.15
django-cors-headers>=4.3
djangorestframework-simplejwt>=5.3
drf-spectacular>=0.27
python-dotenv>=1.0
pytz>=2024.1
jpholiday>=0.1
streamlit>=1.36
pandas>=2.2
altair>=5.3
requests>=2.32
